+++ lib/ruby_aem/handlers/html.rb
@@ -13,13 +13,14 @@
 # limitations under the License.
 
 require 'rexml/document'
-require 'sanitize'
 
 module RubyAem
   # Response handlers for HTML payload.
   # AEM response body needs to be sanitized due to missing closing HTML tags
-  # scattered across many AEM web pages, so for this HTML handler, it whitelists
-  # only the HTML tags that need to be parsed.
+  # scattered across many AEM web pages. The sanitisations are done manually
+  # using simple gsub call in order to avoid dependency to nokogiri which
+  # carries native compilation cost and security vulnerability on libxml
+  # dependency.
   module Handlers
     include REXML
     # Parse authorizable ID from response body data.
@@ -30,7 +31,7 @@ module RubyAem
     # @param call_params API call parameters
     # @return RubyAem::Result
     def self.html_authorizable_id(response, response_spec, call_params)
-      sanitized_body = Sanitize.document(response.body, elements: %w[html title])
+      sanitized_body = _sanitise_html(response.body, /<img.+">/, '')
       html = Document.new(sanitized_body)
       authorizable_id = XPath.first(html, '//title/text()').to_s
       authorizable_id.slice! "Content created #{call_params[:path]}"
@@ -51,8 +52,7 @@ module RubyAem
     # @param call_params API call parameters
     # @return RubyAem::Result
     def self.html_package_service_allow_error(response, response_spec, call_params)
-      sanitized_body = Sanitize.document(response.body, elements: %w[html title p pre])
-      html = Document.new(sanitized_body)
+      html = Document.new(response.body)
       title = XPath.first(html, '//title/text()').to_s
       desc = XPath.first(html, '//p/text()').to_s
       reason = XPath.first(html, '//pre/text()').to_s
@@ -81,7 +81,8 @@ module RubyAem
         raise RubyAem::Error.new(message, result)
       end
 
-      sanitized_body = Sanitize.document(response.body, elements: %w[html body div table tr td b font])
+      sanitized_body = _sanitise_html(response.body, /<input.+>/, '')
+      sanitized_body = _sanitise_html(sanitized_body, /< 0/, '&lt; 0')
       html = Document.new(sanitized_body)
       user = XPath.first(html, '//body/div/table/tr/td/b/text()').to_s
       desc = XPath.first(html, '//body/div/table/tr/td/font/text()').to_s
@@ -94,6 +95,14 @@ module RubyAem
         message = desc
         result = RubyAem::Result.new(message, response)
         raise RubyAem::Error.new(message, result)
+      end
+    end
+
+    def self._sanitise_html(html, regex, replacement)
+      if regex.match?(html)
+        html.gsub!(regex, replacement)
+      else
+        html
       end
     end
   end