+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add new resource aem_configmgr
+
 ## 3.0.0 - 2019-05-14
 ### Added
 - Add application configuration file containing version number