+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Upgrade swagger_aem to 3.1.0
+
 ## 3.1.0 - 2019-06-03
 ### Changed
 - Replace nokogiri with rexml + custom sanitisation for html and xml handlers [#26]
@@ -14,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 3.0.0 - 2019-05-14
 ### Added
 - Add application configuration file containing version number
+
+### Changed
 - Upgrade swagger_aem to 3.0.0
 
 ## 2.5.1 - 2019-02-02